app/requirements-step-2.txt
twilio==6.6.3
ccxt==1.18.361
structlog==17.2.0
python-json-logger==0.1.8
pandas==0.22.0
stockstats==0.2.0
TA-lib==0.4.17
tabulate==0.8.2
slackweb==1.0.5
tenacity==4.8.0
python-telegram-bot==10.0.1
webcord==0.2
jinja2==2.10
requests==2.20.0
PyYAML==3.12
tulipy==0.2.1
matplotlib==3.0.1
scipy==1.1.0
paho-mqtt